backend/queen-ai/core-requirements.txt
# Core Dependencies - OMK Hive Queen AI

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (PostgreSQL)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Cache & Message Queue (Redis)
redis==5.0.1
hiredis==2.2.3  # Redis C parser for better performance

# Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# LLM Providers
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1

# Fetch.ai ASI (Optional - for uAgents)
uagents==0.12.0  # Install separately if needed

# Async & Utils
asyncio==3.4.3
aiohttp==3.9.1
python-dotenv==1.0.0
httpx==0.25.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
tenacity==9.0.0
cachetools<5.0.0,>=4.2.2

# System Management & Monitoring
psutil==5.9.6  # Process and system monitoring
google-cloud-logging==3.8.0  # Cloud Logging (optional)
